refactor: 지역 코드에 id값 추가
