+++ src/apis/index.ts
@@ -2,5 +2,3 @@ export * from '@/apis/aladin';
 export * from '@/apis/groo';
 export * from '@/apis/groo/server';
 export * from '@/apis/library';
-
-export * from '@/apis/library/constants';
