+++ src/types/common/index.ts
@@ -1,0 +1,2 @@
+export * from '@/types/common/dto';
+export * from '@/types/common/region';
