+++ src/types/common/region.ts
@@ -1,21 +1,21 @@
 export const regionList = [
-  { code: '11', name: '서울특별시' },
-  { code: '21', name: '부산광역시' },
-  { code: '22', name: '대구광역시' },
-  { code: '23', name: '인천광역시' },
-  { code: '24', name: '광주광역시' },
-  { code: '25', name: '대전광역시' },
-  { code: '26', name: '울산광역시' },
-  { code: '29', name: '세종특별자치시' },
-  { code: '31', name: '경기도' },
-  { code: '32', name: '강원특별자치도' },
-  { code: '33', name: '충청북도' },
-  { code: '34', name: '충청남도' },
-  { code: '35', name: '전북특별자치도' },
-  { code: '36', name: '전라남도' },
-  { code: '37', name: '경상북도' },
-  { code: '38', name: '경상남도' },
-  { code: '39', name: '제주특별자치도' }
+  { id: 1, code: '11', name: '서울특별시' },
+  { id: 2, code: '21', name: '부산광역시' },
+  { id: 3, code: '22', name: '대구광역시' },
+  { id: 4, code: '23', name: '인천광역시' },
+  { id: 5, code: '24', name: '광주광역시' },
+  { id: 6, code: '25', name: '대전광역시' },
+  { id: 7, code: '26', name: '울산광역시' },
+  { id: 8, code: '29', name: '세종특별자치시' },
+  { id: 9, code: '31', name: '경기도' },
+  { id: 10, code: '32', name: '강원특별자치도' },
+  { id: 11, code: '33', name: '충청북도' },
+  { id: 12, code: '34', name: '충청남도' },
+  { id: 13, code: '35', name: '전북특별자치도' },
+  { id: 14, code: '36', name: '전라남도' },
+  { id: 15, code: '37', name: '경상북도' },
+  { id: 16, code: '38', name: '경상남도' },
+  { id: 17, code: '39', name: '제주특별자치도' }
 ];
 
 export const dtlRegionList = [
