--- src/types/external-api/index.ts
@@ -1,3 +1,0 @@
-// export * from '@/types/external-api/aladin';
-// export * from '@/types/external-api/common';
-// export * from '@/types/external-api/library';
