--- src/types/external-api/library.ts
@@ -1,23 +1,0 @@
-// export interface LibraryBook {
-//   bookname: string;
-//   authors: string;
-//   publisher: string;
-//   publication_year: string;
-//   isbn13: string;
-//   class_no: string;
-//   class_nm: string;
-//   bookImageURL: string;
-//   bookDtlUrl: string;
-// }
-
-// export interface LibrarySearchResponse {
-//   response: {
-//     request: {
-//       keyword: string;
-//       pageNo: number;
-//       pageSize: number;
-//     };
-//     numFound: number;
-//     docs: { doc: LibraryBook }[];
-//   };
-// }
