+++ src/types/index.ts
@@ -1,9 +1,8 @@
-export * from '@/types/common';
-
 export * from '@/types/aladin';
 export * from '@/types/auth';
 export * from '@/types/book';
-// export * from '@/types/external-api';
+export * from '@/types/challenge';
+export * from '@/types/common';
 export * from '@/types/follow';
 export * from '@/types/reviews';
 export * from '@/types/user';
